@@ -1,36 +1,36 @@
 name: unicaen_timetable
 description: Allows the students of the University of Caen Normandie to check their timetable directly from their smartphone.
-version: 3.3.6+87
+version: 3.3.7+88
 
 environment:
   sdk: '>=2.18.0 <3.0.0'
-  flutter: '>=1.10.0'
+  flutter: '>=3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_week_view: ^1.2.2
-  flutter_svg: ^1.1.6
+  flutter_week_view: ^1.3.0
+  flutter_svg: ^2.0.5
   ez_localization: ^1.1.0
   intl: ^0.17.0
-  flutter_secure_storage: ^7.0.1
+  flutter_secure_storage: ^8.0.0
   encrypt: ^5.0.1
-  path_provider: ^2.0.12
+  path_provider: ^2.0.14
   http: ^0.13.5
-  background_fetch: ^1.1.4
+  background_fetch: ^1.1.5
   flutter_colorpicker: ^1.0.3
   share: ^2.0.4
-  rate_my_app: ^1.1.3
-  url_launcher: ^6.1.8
+  rate_my_app: ^1.1.4
+  url_launcher: ^6.1.10
   package_info: ^2.0.2
   device_info: ^2.0.3
   scrollable_positioned_list: ^0.3.5
-  collection: ^1.17.0
-  google_mobile_ads: ^2.3.0
-  flutter_riverpod: ^2.1.3
+  google_mobile_ads: ^2.4.0
+  flutter_riverpod: ^2.3.4
   user_messaging_platform: ^1.2.1
-  flutter_widget_from_html_core: ^0.9.1
-  shared_preferences: ^2.0.17
+  flutter_widget_from_html_core: ^0.10.0
+  shared_preferences: ^2.1.0
+  catcher: ^0.7.0
 
 dev_dependencies:
   flutter_lints: ^2.0.1
