@@ -95,19 +95,19 @@ abstract class UnicaenTimetableTheme {
           color: actionBarColor,
         ),
         textTheme: TextTheme(
-          headline1: TextStyle(color: textColor),
-          headline2: TextStyle(color: textColor),
-          headline3: TextStyle(color: textColor),
-          headline4: TextStyle(color: textColor),
-          headline5: TextStyle(color: textColor),
-          headline6: TextStyle(color: textColor),
-          subtitle1: TextStyle(color: textColor),
-          subtitle2: TextStyle(color: textColor),
-          bodyText1: TextStyle(color: textColor),
-          bodyText2: TextStyle(color: textColor),
-          caption: TextStyle(color: textColor),
+          displayLarge: TextStyle(color: textColor),
+          displayMedium: TextStyle(color: textColor),
+          displaySmall: TextStyle(color: textColor),
+          headlineMedium: TextStyle(color: textColor),
+          headlineSmall: TextStyle(color: textColor),
+          titleLarge: TextStyle(color: textColor),
+          titleMedium: TextStyle(color: textColor),
+          titleSmall: TextStyle(color: textColor),
+          bodyLarge: TextStyle(color: textColor),
+          bodyMedium: TextStyle(color: textColor),
+          bodySmall: TextStyle(color: textColor),
           // button: TextStyle(color: textColor),
-          overline: TextStyle(color: textColor),
+          labelSmall: TextStyle(color: textColor),
         ),
         listTileTheme: ListTileThemeData(
           textColor: listTileTextColor,
